backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-pm-saas-backend"
version = "0.1.0"
description = "Backend API for AI-Powered Project Management SaaS"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.34.0",
  "pydantic==2.10.5",
  "pydantic-settings==2.7.1",
  "motor==3.6.0",
  "redis==5.2.1",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==4.1.3",
  "python-multipart==0.0.20",
  "email-validator==2.2.0",
  "celery==5.4.0",
  "httpx==0.28.1",
  "slowapi==0.1.9",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "pytest-asyncio==0.25.2",
  "pytest-cov==6.0.0",
  "ruff",
  "mypy",
  "black",
  "isort",
  "bandit",
  "safety",
  "pip-tools",
  "pip-audit",
  "pip-licenses",
  "radon",
]

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
